+++ src/samplers/GradientDescent.py
@@ -35,9 +35,9 @@ class GradientDescent(tf.train.GradientDescentOptimizer):
         lr_t = math_ops.cast(self._learning_rate_t, var.dtype.base_dtype)
         scaled_gradient = lr_t * grad
         with tf.variable_scope("accumulate", reuse=True):
-            gradient_global = tf.get_variable("gradients")
+            gradient_global = tf.get_variable("gradients", dtype=tf.float64)
             gradient_global_t = tf.assign_add(gradient_global, tf.reduce_sum(tf.multiply(scaled_gradient, scaled_gradient)))
-            virial_global = tf.get_variable("virials")
+            virial_global = tf.get_variable("virials", dtype=tf.float64)
             virial_global_t = tf.assign_add(virial_global, tf.reduce_sum(tf.multiply(grad, var)))
         control_group_gradient_descent_t = super(GradientDescent, self)._apply_dense(grad, var)
         return control_flow_ops.group(*[virial_global_t, control_group_gradient_descent_t, gradient_global_t])
